refactor: move `accounts` to standard path
